Refactor testcases
