+++ api/tests/oauth_app_response.py
@@ -1,0 +1,24 @@
+
+
+GITHUB_USER_RESPONSE = {
+    "id": 14,
+    "login": "amazingguni",
+    "avatar_url": "https://avatars0.githubusercontent.com/u/13607839?v=4",
+    "type": "User",
+    "name": "amazingguni_test",
+    "location": "Seoul, Korea",
+    "email": "amazingguni@gmail.com",
+    "bio": "Be a good engineer",
+}
+
+GITHUB_EMAILS_RESPONSE = [
+    {
+        "email": "amazingguni@pycon.kr",
+        "primary": True,
+        "verified": True,
+        "visibility": "public"
+    }
+]
+
+GITHUB_ACCESS_TOKEN_RESPONSE = \
+    b'access_token=111767d60d69f06812b54ca7b5bb46c49358a4ec&token_type=bearer'
